refactor(common): lift git IPC payload types out of @beak/state (ADR 0003)
Moved seven Redux action-payload types from @beak/state/src/git/index.ts
into @beak/common/src/ipc/git.ts, where the rest of the git IPC contract
already lives:

  GitInitRequest, GitCommitRequest, GitPushRequest, GitPullRequest,
  GitFetchRequest, GitCheckoutRequest, GitCreateBranchRequest

Per ADR 0003 §2: types imported by an IPC handler or that cross the
renderer↔host boundary belong in @beak/common, not @beak/state.  These
types drive what the renderer asks the host to do (the host already
validated them via the @beak/common Zod schemas), so they are IPC
contract types, not slice-internal state.

@beak/state/git re-exports all seven types from @beak/common for
backward compatibility; existing importers in @beak/ui are unaffected.
Both host git-service files already imported exclusively from
@beak/common and required no changes.
